Run sqlite optimization after ingestion db connection closes:
This is recommended practice for SQLite databases with short-lived
connections such as the connection for ingestion:

  https://sqlite.org/pragma.html#pragma_optimize

Note the callout of "low-quality indexes" on this page:

  https://sqlite.org/queryplanner-ng.html#howtofix

Due to the nature of both topics and contract IDs, we have a huge
variance in the K -> V ratio. Things like XLM have tons of event
rows given fee events occur on every txn, and "transfer" is
obviously a popular topic1 symbol. This explicitly breaks an
indexing rule:

> A low-quality index is one where there are **more than 10 or 20
> rows in the table that have the same value** for the
> left-most column of the index.

The solution is running `ANALYZE events;`, but this would take too
long and grabs an exclusive lock on the DB, so we must periodically
run `ANALYZE` via the `PRAGMA optimize;` directive which we
execute at the end of every ingestion commit.